Move app e2e smoke data out of spec
