+++ test/e2e/app.spec.ts
@@ -4,269 +4,18 @@
  * Run with:
  *   yarn test:e2e:app
  */
-import { test, expect, type Page } from "@playwright/test";
-import type { MoreInfoView } from "../../src/dialogs/more-info/const";
+import { test, expect } from "@playwright/test";
 import {
+  defineLinkSmokeTests,
   defineRouteSmokeTests,
   goToPanel,
   PANEL_TIMEOUT,
   QUICK_TIMEOUT,
-  rendersRoute,
-  routeCase,
-  routeCases,
   SHELL_TIMEOUT,
-  type RouteSmokeCase,
-  type RouteSmokeGroup,
 } from "./helpers";
-import { e2ePanelRouteAssertions } from "./app/src/ha-test-panels";
-
-/**
- * Each More info view renders one root element inside the dialog, plus one or
- * more characteristic descendants that prove the view actually populated rather
- * than rendering an empty shell. `text`, when set, asserts the element's text
- * instead of just its presence.
- */
-const MORE_INFO_VIEW_ELEMENTS: {
-  view: MoreInfoView;
-  element: string;
-  content: { selector: string; text?: string }[];
-}[] = [
-  {
-    view: "info",
-    element: "ha-more-info-info",
-    content: [
-      { selector: "more-info-light" },
-      { selector: "span.title", text: "Test Light" },
-    ],
-  },
-  {
-    view: "history",
-    element: "ha-more-info-history-and-logbook",
-    // The demo loads the history component but not logbook.
-    content: [{ selector: "ha-more-info-history" }],
-  },
-  {
-    view: "settings",
-    element: "ha-more-info-settings",
-    // The scenario mocks config/entity_registry/get, so the real registry
-    // panel renders instead of the "no unique ID" warning.
-    content: [{ selector: "entity-registry-settings" }],
-  },
-  {
-    view: "related",
-    element: "ha-related-items",
-    // search/related is mocked to return no relations, so the empty list
-    // renders.
-    content: [{ selector: "ha-related-items >> ha-list" }],
-  },
-  {
-    view: "add_to",
-    element: "ha-more-info-add-to",
-    // Admin users get the default add-to action list.
-    content: [{ selector: "ha-add-to-action-list" }],
-  },
-  {
-    view: "details",
-    element: "ha-more-info-details",
-    // The details view renders the state and attributes cards.
-    content: [{ selector: "ha-card" }],
-  },
-];
-
-interface E2ELovelaceRoot extends HTMLElement {
-  lovelace?: {
-    setEditMode: (editMode: boolean) => void;
-  };
-}
-
-async function setLovelaceEditMode(page: Page, editMode: boolean) {
-  await page
-    .locator("hui-root")
-    .first()
-    .waitFor({ state: "attached", timeout: QUICK_TIMEOUT });
-  await page
-    .locator("hui-root")
-    .first()
-    .evaluate(async (el: Element, value) => {
-      const root = el as E2ELovelaceRoot;
-      const start = performance.now();
-      await new Promise<void>((resolve, reject) => {
-        const check = () => {
-          if (root.lovelace?.setEditMode) {
-            resolve();
-            return;
-          }
-          if (performance.now() - start > 2000) {
-            reject(new Error("Lovelace edit mode action was not available"));
-            return;
-          }
-          requestAnimationFrame(check);
-        };
-        check();
-      });
-      root.lovelace!.setEditMode(value);
-    }, editMode);
-}
-
-const PANEL_ROUTE_ASSERTIONS: RouteSmokeCase[] = Array.from(
-  e2ePanelRouteAssertions,
-  ([path, element]) => routeCase(path, element)
-);
-
-const URL_NORMALIZATION_ASSERTIONS: RouteSmokeCase[] = [
-  {
-    name: "keeps the todo panel when adding the selected entity query",
-    path: "/todo",
-    element: "ha-panel-todo",
-    url: /\/\?entity_id=todo\.shopping_list#\/todo$/,
-  },
-  {
-    name: "keeps the history panel when removing the back query",
-    path: "/?back=1#/history",
-    element: "ha-panel-history, history-panel",
-    url: /\/#\/history$/,
-  },
-  {
-    name: "keeps the logbook panel when removing the back query",
-    path: "/?back=1#/logbook",
-    element: "ha-panel-logbook",
-    url: /\/#\/logbook$/,
-  },
-  {
-    name: "keeps the lovelace panel when adding the edit query",
-    path: "/lovelace",
-    element: "ha-panel-lovelace, hui-root",
-    url: /\/\?edit=1#\/lovelace\/home$/,
-    action: (page) => setLovelaceEditMode(page, true),
-  },
-  {
-    name: "keeps the lovelace panel when removing the edit query",
-    path: "/lovelace",
-    element: "ha-panel-lovelace, hui-root",
-    url: /\/#\/lovelace\/home$/,
-    action: async (page) => {
-      await setLovelaceEditMode(page, true);
-      await expect(page).toHaveURL(/\/\?edit=1#\/lovelace\/home$/, {
-        timeout: SHELL_TIMEOUT,
-      });
-      await setLovelaceEditMode(page, false);
-    },
-  },
-];
-
-const TOOLS_SUBPAGES: { route: string; element: string }[] = [
-  { route: "yaml", element: "tools-yaml-config" },
-  { route: "state", element: "tools-state" },
-  { route: "action", element: "tools-action" },
-  { route: "template", element: "tools-template" },
-  { route: "event", element: "tools-event" },
-  { route: "statistics", element: "tools-statistics" },
-  { route: "assist", element: "tools-assist" },
-  { route: "debug", element: "tools-debug" },
-];
-
-const TOOLS_ROUTE_ASSERTIONS: RouteSmokeCase[] = [
-  routeCase("/config/tools", "ha-panel-tools"),
-  ...TOOLS_SUBPAGES.map(({ route, element }) =>
-    routeCase(`/config/tools/${route}`, element)
-  ),
-  routeCase("/config/tools/service", "tools-action"),
-];
-
-const TOOLS_REDIRECT_ASSERTIONS: RouteSmokeCase[] = [
-  ...["/developer-tools", "/config/developer-tools"].flatMap((oldBase) => [
-    routeCase(oldBase, "ha-panel-tools"),
-    routeCase(`${oldBase}/state`, "tools-state"),
-  ]),
-];
-
-const CONFIG_ROUTES = routeCases([
-  ["/config/integrations", "ha-config-integrations"],
-  ["/config/devices", "ha-config-devices"],
-  ["/config/entities", "ha-config-entities"],
-  ["/config/helpers", "ha-config-helpers"],
-  ["/config/areas", "ha-config-areas"],
-  ["/config/apps", "ha-config-apps"],
-  ["/config/app", "ha-config-app-dashboard"],
-  ["/config/automation", "ha-config-automation"],
-  ["/config/backup", "ha-config-backup"],
-  ["/config/scene", "ha-config-scene"],
-  ["/config/script", "ha-config-script"],
-  ["/config/blueprint", "ha-config-blueprint"],
-  ["/config/cloud", "ha-config-cloud"],
-  ["/config/energy", "ha-config-energy"],
-  ["/config/hardware", "ha-config-hardware"],
-  ["/config/labs", "ha-config-labs"],
-  ["/config/lovelace", "ha-config-lovelace"],
-  ["/config/person", "ha-config-person"],
-  ["/config/storage", "ha-config-section-storage"],
-  ["/config/tags", "ha-config-tags"],
-  ["/config/users", "ha-config-users"],
-  ["/config/voice-assistants", "ha-config-voice-assistants"],
-  ["/config/system", "ha-config-system-navigation"],
-  ["/config/info", "ha-config-info"],
-  ["/config/logs", "ha-config-logs"],
-  ["/config/general", "ha-config-section-general"],
-  ["/config/updates", "ha-config-section-updates"],
-  ["/config/repairs", "ha-config-repairs-dashboard"],
-  ["/config/analytics", "ha-config-section-analytics"],
-  ["/config/ai-tasks", "ha-config-section-ai-tasks"],
-  ["/config/labels", "ha-config-labels"],
-  ["/config/zone", "ha-config-zone"],
-  ["/config/network", "ha-config-section-network"],
-  ["/config/application_credentials", "ha-config-application-credentials"],
-  ["/config/bluetooth", "bluetooth-config-dashboard-router"],
-  ["/config/dhcp", "dhcp-config-panel"],
-  ["/config/infrared", "infrared-config-dashboard-router"],
-  ["/config/matter", "matter-config-panel"],
-  ["/config/mqtt", "mqtt-config-panel"],
-  ["/config/radio-frequency", "radio-frequency-config-dashboard-router"],
-  ["/config/ssdp", "ssdp-config-panel"],
-  ["/config/thread", "thread-config-panel"],
-  ["/config/zeroconf", "zeroconf-config-panel"],
-  ["/config/zha", "zha-config-dashboard-router"],
-  ["/config/zwave_js", "zwave_js-config-router"],
-]);
-
-const NESTED_CONFIG_ROUTES = routeCases([
-  ["/config/integrations/dashboard", "ha-config-integrations-dashboard"],
-  ["/config/devices/dashboard", "ha-config-devices-dashboard"],
-  ["/config/areas/dashboard", "ha-config-areas-dashboard"],
-  ["/config/backup/settings", "ha-config-backup-settings"],
-]);
-
-const ROUTE_SMOKE_GROUPS: RouteSmokeGroup[] = [
-  {
-    name: "Panel navigation",
-    routes: PANEL_ROUTE_ASSERTIONS,
-    testName: (route) => `renders registered panel ${route.path}`,
-  },
-  {
-    name: "Panel URL normalization",
-    routes: URL_NORMALIZATION_ASSERTIONS,
-    testName: (route) => route.name!,
-  },
-  {
-    name: "Tools panel",
-    routes: TOOLS_ROUTE_ASSERTIONS,
-    testName: rendersRoute,
-  },
-  {
-    name: "Tools redirects",
-    routes: TOOLS_REDIRECT_ASSERTIONS,
-    testName: (route) => `redirects ${route.path}`,
-  },
-  {
-    name: "Config routes",
-    routes: CONFIG_ROUTES,
-    testName: rendersRoute,
-  },
-  {
-    name: "Nested config routes",
-    routes: NESTED_CONFIG_ROUTES,
-    testName: rendersRoute,
-  },
-];
+import { configLinks } from "./app/src/config-smoke";
+import { moreInfoViewElements } from "./app/src/more-info-smoke";
+import { appRouteSmokeGroups } from "./app/src/route-smoke";
 
 // ---------------------------------------------------------------------------
 // App shell
@@ -388,7 +137,7 @@ test.describe("App shell", () => {
   });
 });
 
-defineRouteSmokeTests(ROUTE_SMOKE_GROUPS);
+defineRouteSmokeTests(appRouteSmokeGroups);
 
 // ---------------------------------------------------------------------------
 // Lovelace
@@ -417,7 +166,7 @@ test.describe("Lovelace dashboard", () => {
 // ---------------------------------------------------------------------------
 
 test.describe("Light more-info dialog", () => {
-  for (const { view, element, content } of MORE_INFO_VIEW_ELEMENTS) {
+  for (const { view, element, content } of moreInfoViewElements) {
     test(`opens more-info ${view} view for a light entity`, async ({
       page,
     }) => {
@@ -521,29 +270,6 @@ test.describe("Theming", () => {
 // ---------------------------------------------------------------------------
 
 test.describe("Config panel", () => {
-  const DASHBOARD_LINKS = [
-    { href: "/config/integrations", label: "Devices & services" },
-    { href: "/config/automation", label: "Automations & scenes" },
-    { href: "/config/areas", label: "Areas, labels & zones" },
-    { href: "/config/apps", label: "Apps" },
-    { href: "/config/lovelace/dashboards", label: "Dashboards" },
-    { href: "/config/voice-assistants", label: "Voice assistants" },
-    { href: "/config/matter", label: "Matter" },
-    { href: "/config/zha", label: "Zigbee" },
-    { href: "/config/zwave_js", label: "Z-Wave" },
-    { href: "/knx", label: "KNX" },
-    { href: "/config/thread", label: "Thread" },
-    { href: "/config/bluetooth", label: "Bluetooth" },
-    { href: "/config/infrared", label: "Infrared" },
-    { href: "/config/radio-frequency", label: "Radio frequency" },
-    { href: "/insteon", label: "Insteon" },
-    { href: "/config/tags", label: "Tags" },
-    { href: "/config/person", label: "People" },
-    { href: "/config/system", label: "System" },
-    { href: "/config/tools", label: "Tools" },
-    { href: "/config/info", label: "About" },
-  ];
-
   test("config panel loads without JS errors", async ({ page }) => {
     const errors: string[] = [];
     page.on("pageerror", (e) => errors.push(e.message));
@@ -560,20 +286,12 @@ test.describe("Config panel", () => {
     expect(realErrors).toHaveLength(0);
   });
 
-  test("dashboard renders key settings links", async ({ page }) => {
+  const getDashboard = async (page) => {
     await goToPanel(page, "/config");
-
     const dashboard = page.locator("ha-config-dashboard");
     await expect(dashboard).toBeAttached({ timeout: PANEL_TIMEOUT });
+    return dashboard;
+  };
 
-    for (const { href, label } of DASHBOARD_LINKS) {
-      const link = dashboard.getByRole("link", {
-        name: new RegExp(`^${label.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}\\b`),
-      });
-      // eslint-disable-next-line no-await-in-loop
-      await expect(link).toHaveAttribute("href", href, {
-        timeout: QUICK_TIMEOUT,
-      });
-    }
-  });
+  defineLinkSmokeTests(configLinks, getDashboard);
 });
